added LabelController tests

+++ resources/views/app/statuses/show.blade.php
@@ -1,8 +1,8 @@
 @extends('layouts.app')
 @section('content')
     <h1 class="mb-5">Statuses</h1>
-    @if(Auth::check())
-    <a class="btn btn-primary" href="{{route('statuses.create')}}">{{ __('Create Status') }}</a>
+    @if(auth()->user())
+    <a class="btn btn-primary" href="{{route('statuses.create')}}">{{ __('Create status') }}</a>
     @endif
     <table class="table mt-2">
         <thead>
@@ -20,15 +20,17 @@
                     <td>{{$status->name}}</td>
                     <td>{{$status->created_at}}</td>
                     <td>
-                        @if(Auth::check())
-                        <a class="btn btn-danger btn-sm"
-                           data-confirm="Вы уверены?" data-method="delete" rel="nofollow" href="{{ route('statuses.destroy', $status) }}">
-                            {{__('delete')}}
-                        </a>
-                        <a class="btn btn-success btn-sm" href="{{ route('statuses.edit', $status) }}">
-                            {{__('edit')}}
-                        </a>
-                        @endif
+                            @can('delete', $status)
+                                <a class="btn btn-danger btn-sm ml-2"
+                                data-confirm="Вы уверены?" data-method="delete" rel="nofollow" href="{{ route('statuses.destroy', $status) }}">
+                                    {{__('delete')}}
+                                </a>
+                            @endcan
+                            @can('update', $status)
+                                <a class="btn btn-success btn-sm float-left" href="{{ route('statuses.edit', $status) }}">
+                                    {{__('edit')}}
+                                </a>
+                            @endcan
                     </td>
                 </tr>
             @endforeach
